+++ source_gen/source_gen/source_gen/test/library/find_type_test.dart
@@ -15,14 +15,17 @@ final _source = r'''
   import 'dart:async' show Stream;
 
   part 'part.dart';
-  
+
   class Example {}
+  enum Enum{A,B}
 ''';
 
 final _partSource = r'''
 part of 'source.dart';
 
 class PartClass {}
+
+enum PartEnum{A,B}
 ''';
 
 void main() {
@@ -36,7 +39,13 @@ void main() {
   });
 
   test('class count', () {
-    expect(library.classElements, hasLength(2));
+    expect(library.classes.map((c) => c.name), ['Example', 'PartClass']);
+    // ignore: deprecated_member_use
+    expect(library.classElements, orderedEquals(library.classes));
+  });
+
+  test('enum count', () {
+    expect(library.enums.map((e) => e.name), ['Enum', 'PartEnum']);
   });
 
   test('should return a type not exported', () {